Bug #1860: Specified fabs(double) to resolve ambiguous call on Windows

+++ Code/Applications/niftkExtractScalp.cxx
@@ -128,8 +128,8 @@ int DoMain(arguments args)
     maskPixel = maskIterator.Get();
     imagePixel = imageIterator.Get();
 
-    if (fabs(imagePixel - 0) > args.tolerance      // Non-zero voxels
-        && fabs(maskPixel - 0) < args.tolerance)   // That are non-brain
+    if (fabs((double)(imagePixel - 0)) > args.tolerance      // Non-zero voxels
+        && fabs((double)(maskPixel - 0)) < args.tolerance)   // That are non-brain
     {
       voxelCounter++;
       t_skull += imagePixel;
@@ -150,7 +150,7 @@ int DoMain(arguments args)
     maskPixel = maskIterator.Get();
     imagePixel = imageIterator.Get();
 
-    if (fabs(maskPixel - 0) < args.tolerance  // non-brain voxels
+    if (fabs((double)(maskPixel - 0)) < args.tolerance  // non-brain voxels
         && imagePixel >= t_skull)
     {
       voxelCounter++;
